requirements.txt
tensorflow>=2.10
numpy
opencv-python
matplotlib
scikit-learn
pillow
tqdm
kaggle
flask
flask-cors
pyserial
